// src/types.ts
export interface Estudiante {
  id: string;
  nombreCompleto: string;
  curso?: string | null;
}